refactor: move preview utils in a separated file
